Starting on sequences, moved `LiveActor` to `liveActor`
