+++ lib/al/liveActor/ShadowKeeper.h
