+++ lib/al/liveActor/ShadowMaskCtrl.h
