+++ lib/al/player/PlayerHackKeeper.h
@@ -5,7 +5,7 @@
 
 #pragma once
 
-#include "al/LiveActor/LiveActor.h"
+#include "al/liveActor/LiveActor.h"
 
 class PlayerHackKeeper {
 public:
