+++ lib/al/scene/SceneCreator.h
@@ -1,0 +1,5 @@
+#pragma once
+
+namespace al {
+class SceneCreator {};
+}  // namespace al
