+++ lib/game/sequence/E3Sequence.h
@@ -1,0 +1,3 @@
+#pragma once
+
+class E3Sequence {};
