+++ lib/game/sequence/HakoniwaSequence.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include "al/sequence/Sequence.h"
+
+class HakoniwaSequence : al::Sequence {
+
+};
