+++ src/al/liveActor/ShadowKeeper.cpp
