+++ src/game/scene/CMakeLists.txt
@@ -1,0 +1,3 @@
+target_sources(odyssey PRIVATE
+    StageScene.cpp
+)
