+++ src/game/scene/StageScene.cpp
@@ -1,0 +1,1 @@
+#include "game/scene/StageScene.h"
