+++ src/game/sequence/CMakeLists.txt
@@ -1,0 +1,4 @@
+target_sources(odyssey PRIVATE
+    E3Sequence.cpp
+    HakoniwaSequence.cpp
+)
