+++ src/game/sequence/HakoniwaSequence.cpp
@@ -1,0 +1,1 @@
+#include "game/sequence/HakoniwaSequence.h"
